clean code and separate trajectories for train and val

+++ make_dataset.py
@@ -139,7 +139,7 @@ if __name__ == '__main__':
     trajectory_rewards = trajectory_rewards[p]
 
     # Split
-    split_i = np.ceil(val_split)
+    split_i = int(np.ceil(val_split))
     train_trajectories = trajectories[0:split_i]
     train_trajectory_rewards = trajectory_rewards[0:split_i]
     val_trajectories = trajectories[split_i:]
